pyproject.toml
[project]
name = "iconconverter"
version = "1.4.0"
description = "Image to ICO converter with transparency support"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "IconConverter Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
iconconverter = "iconconverter.gui:main"

[project.urls]
Homepage = "https://github.com/yourusername/iconconverter"
Repository = "https://github.com/yourusername/iconconverter.git"
Issues = "https://github.com/yourusername/iconconverter/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C90", # mccabe complexity
    "T20", # flake8-print (禁止print検出)
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T20"]  # テストではprint許可

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
disallow_any_generics = true
warn_return_any = true
strict_equality = true
show_error_codes = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["iconconverter"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true
fail_under = 80  # production段階では80%

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "safety>=3.6.2",
]
